Pulling Toolbar into it's own Component
