+++ src/app/shared/top-toolbar/top-toolbar.component.css
@@ -1,0 +1,23 @@
+.title-toolbar {
+  --mat-toolbar-standard-height: 4em;
+
+  background-color: var(--color-squid-mid2green);
+  color: var(--color-dark-foreground);
+  font-family: Metropolis, sans-serif;
+
+  .menu-icon {
+    color: var(--color-dark-foreground);
+  }
+  .menu-spacer {
+    flex: 1 1 auto;
+    text-align: center;
+    position: relative;
+
+  }
+
+  .menu-logo {
+    position: relative;
+    height: 3.5em;
+    top: 0.25em;
+  }
+}
