+++ src/app/shared/top-toolbar/top-toolbar.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ToolbarService } from './toolbar.service';
+import { TopToolbarService } from './top-toolbar.service';
 
-describe('ToolbarService', () => {
-  let service: ToolbarService;
+describe('TopToolbarService', () => {
+  let service: TopToolbarService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(ToolbarService);
+    service = TestBed.inject(TopToolbarService);
   });
 
   it('should be created', () => {
